Improve classifier pipeline, add weight generation
Continue building a pipeline add funcitonalities to weight saving,
parameters can be saved in files. Weights are written with certain format
in files ALGORITHM_DATASET_Description, ALGORITHM_DATASET_1, ...
generated files are in output/dataset directory

+++ uvnn/utils/preprocessors.py
@@ -19,7 +19,7 @@ class BasicPreprocessor(object):
     def preprocess_data(self):
         ''' Basic preprocessing normalizes the data, substracts mean
         and divides by sd '''
-        self.y = self.y.flatten()
+        self.y = self.y.flatten().astype(int)
         if self.hook is not None:
             self.X, self.y = self.hook(self.X, self.y)
